requirements.txt
Jinja2
PyYAML
click